+++ lib/Resque/Resque.php
@@ -14,7 +14,7 @@ class Resque
     const DEFAULT_INTERVAL = 5;
 
     /**
-     * @var Resque_Redis Instance of Resque_Redis that talks to redis.
+     * @var Resque\Redis Instance of Resque\Redis that talks to redis.
      */
     public static $redis = null;
 
@@ -35,7 +35,7 @@ class Resque
      *
      * @param mixed $server Host/port combination separated by a colon, DSN-formatted URI, or
      *                      a callable that receives the configured database ID
-     *                      and returns a Resque_Redis instance, or
+     *                      and returns a Resque\Redis instance, or
      *                      a nested array of servers with host/port pairs.
      * @param int $database
      */
@@ -47,9 +47,9 @@ class Resque
     }
 
     /**
-     * Return an instance of the Resque_Redis class instantiated for Resque.
+     * Return an instance of the Resque\Redis class instantiated for Resque.
      *
-     * @return Resque_Redis Instance of Resque_Redis.
+     * @return Resque\Redis Instance of Resque\Redis.
      */
     public static function redis()
     {
@@ -60,7 +60,7 @@ class Resque
         if (is_callable(self::$redisServer)) {
             self::$redis = call_user_func(self::$redisServer, self::$redisDatabase);
         } else {
-            self::$redis = new \Resque_Redis(self::$redisServer, self::$redisDatabase);
+            self::$redis = new Redis(self::$redisServer, self::$redisDatabase);
         }
 
         return self::$redis;
@@ -179,7 +179,7 @@ class Resque
         }
 
         /**
-         * Normally the Resque_Redis class returns queue names without the prefix
+         * Normally the Resque\Redis class returns queue names without the prefix
          * But the blpop is a bit different. It returns the name as prefix:queue:name
          * So we need to strip off the prefix:queue: part
          */